+++ relay-dynamic-config/src/defaults.rs
@@ -118,6 +118,20 @@ pub fn add_span_metrics(project_config: &mut ProjectConfig) {
             ),
             tags: Default::default(),
         },
+        MetricSpec {
+            category: DataCategory::Span,
+            mri: "c:spans/count_per_op@none".into(),
+            field: None,
+            condition: Some(duration_condition.clone()),
+            tags: ["category", "op"]
+                .map(|key| TagSpec {
+                    key: format!("span.{key}"),
+                    field: Some(format!("span.sentry_tags.{key}")),
+                    value: None,
+                    condition: None,
+                })
+                .into(),
+        },
     ]);
 
     config.tags.extend([